change to make the hook its own file
